drones-simulator/requirements.txt
pymongo==4.6.1
Flask==3.0.1
kafka-python==2.0.2
numpy==1.26.3